создал модуляцию, разбил по файлам
