+++ js/data.js
@@ -1,0 +1,75 @@
+// массив аватарок
+const AVATARS = [];
+
+// массив тайтлов
+const OFFER_TITLES = [
+  'Bisness',
+  'Cozy',
+  'Beautiful',
+  'Luxary',
+  'Nice',
+  'Georges',
+  'Stylish',
+  'Modern',
+  'Contemporary',
+  'New',
+  'Comfortable',
+  'Great',
+];
+
+//тип жилья
+const OFFER_TYPE = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+
+// время поселения
+const OFFER_CHECKIN = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+// время выселения
+const OFFER_CHECKOUT = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+// дополнительные опции
+const OFFER_FEAUTURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+// описание квартиры
+const OFFER_DESCRIPTIONS = [
+  'The living is easy in this impressive, generously proportioned contemporary residence with lake and ocean views, located within a level stroll to the sand and surf.',
+  'This immaculately presented apartment is set amongst manicured grounds within a private and secure complex. As a resident, you will have access to lifestyle amenities including a lap pool, gymnasium, communal terraces, concierge service and basement parking.',
+  'Prepare to be impressed when you enter this superbly maintained and presented home set on a sprawling 1/2 acre parcel of land occupying a peaceful street position.',
+  'Cool, calm and sophisticated with a youthful edge, this functional home is enveloped in light and comfort. Crisp white walls, timber floors, and high ceilings create a style as timeless as the sparkling ocean view. The calming sea vista, captured through the extensive use of glass, will help you forget your city stress.',
+  'Embrace the spirit of DIY with this original cottage occupying a peaceful street position. This charming weatherboard home features 3 roomy bedrooms and bright, open living spaces.',
+  'Magnificent ocean and beach views consume the top floor of this meticulously finished and modern property. Bi-fold doors create a seamless flow from the vast dining, kitchen and living areas to the expansive deck perfect for entertaining.',
+  'The house comes complete with two living rooms, a welcoming kitchen/dining area, two bathrooms, four bedrooms, a study, and a laundry, and retains the value of peaceful living while being conveniently close to shops, school and transport.',
+  'Soaring ceilings, white walls and timber floors add a dramatic tone to the all-encompassing view. Enjoy the comfort of a cool, shady cabana between dips in the saltwater pool or dive into the surf, just a few sandy steps from your front door.',
+  'This house screams ‘designer’ and will reflect the personality and taste of those accustomed to the best in quality design, finishes and lifestyle.',
+  'This gorgeous home provides lifestyle options aplenty: kick off your shoes for a bit of in-home relaxation, or kick off your heels and party into the night in the city’s vibrant entertainment precinct – only 20 minutes away by road or rail.',
+];
+
+// фотографии квартиры
+const OFFER_IMGS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+
+export {AVATARS, OFFER_TITLES, OFFER_TYPE, OFFER_CHECKIN, OFFER_CHECKOUT, OFFER_FEAUTURES, OFFER_DESCRIPTIONS, OFFER_IMGS};
